scrapers/requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.3
python-dateutil>=2.8.2
icalendar>=5.0.11
feedparser>=6.0.11
extruct>=0.17.0
w3lib>=2.1.2
pytz>=2024.1
